[FEATURE] Voter import
To fill the voter directory, this commit provides a CLI controller and a
sophisticated class structure for describing the import format. The
import can be done from any file where each record has its own line, the
fields can be separated arbitrarily (see the example in
Configuration/Settings.yaml for more information).

The import has been tested on a dataset of 25.000 records and except a
rather high memory usage, no difficulties have been found. This could
perhaps be reduced by dividing the import into smaller chunks of e.g.
1000 records, though this would make it harder to have the import as one
single transaction.

Still missing is a sophisticated error handling, e.g. to ensure
uniqueness of discriminator values like the matriculation number. This
should normally be no concern if the input data is well-defined, but as
the correctness of the election depends on this, we should check it
nonetheless.