[FEATURE] Search voter directory
The voter directory can now be searched, currently only using the name
of the voter, not the discriminators.

The search is implemented as its own class to pass the parameters around
and encapsulate the search process/results. The object currently cannot
be passed around really well between actions, as Flow does not support
passing regular, non-persisted objects.